Add a base ball-ball collision substrate
