+++ pooltool/physics/resolve/ball_ball/frictionless_elastic/__init__.py
@@ -5,6 +5,7 @@ import numpy as np
 import pooltool.constants as const
 import pooltool.math as math
 from pooltool.objects.ball.datatypes import Ball, BallState
+from pooltool.physics.resolve.ball_ball.core import CoreBallBallCollision
 
 
 def _resolve_ball_ball(rvw1, rvw2, R):
@@ -27,16 +28,8 @@ def _resolve_ball_ball(rvw1, rvw2, R):
     return rvw1, rvw2
 
 
-class FrictionlessElastic:
-    def resolve(
-        self, ball1: Ball, ball2: Ball, inplace: bool = False
-    ) -> Tuple[Ball, Ball]:
-        if not inplace:
-            ball1 = ball1.copy()
-            ball2 = ball2.copy()
-
-        ball1, ball2 = self.make_kiss(ball1, ball2)
-
+class FrictionlessElastic(CoreBallBallCollision):
+    def solve(self, ball1: Ball, ball2: Ball) -> Tuple[Ball, Ball]:
         rvw1, rvw2 = _resolve_ball_ball(
             ball1.state.rvw.copy(),
             ball2.state.rvw.copy(),
@@ -45,22 +38,5 @@ class FrictionlessElastic:
 
         ball1.state = BallState(rvw1, const.sliding)
         ball2.state = BallState(rvw2, const.sliding)
-
-        return ball1, ball2
-
-    def make_kiss(self, ball1: Ball, ball2: Ball) -> Tuple[Ball, Ball]:
-        """Translate the balls so they are (almost) touching
-
-        This makes a correction such that if the balls are not 2*R apart, they are moved
-        equally along their line of centers such that they are. To avoid float precision
-        round-off error, a small epsilon of additional distance (constants.EPS_SPACE) is
-        put between them, ensuring the balls are non-intersecting.
-        """
-        r1, r2 = ball1.state.rvw[0], ball2.state.rvw[0]
-        n = math.unit_vector(r2 - r1)
-
-        correction = 2 * ball1.params.R - math.norm3d(r2 - r1) + const.EPS_SPACE
-        ball2.state.rvw[0] += correction / 2 * n
-        ball1.state.rvw[0] -= correction / 2 * n
 
         return ball1, ball2
